During selfupdate-finsh, ensure fink-bld user
With --build-as-nobody soon to be default, we want to make ensure that
the fink-bld user is available for use during building. This patch adds
a fink-bld user and group if it is not installed during
selfupdate-finish. Unlike the passwd package, the uid and gid are not
fixed, but are chose not to conflict with any current users.